fix(models): quote ordering identifiers and cast product ids (fixes #2526)
Admin list models built their ORDER BY column with the value escaper
rather than the identifier quoter, so the clause read as though escape()
were what constrained it. The column is in fact resolved by
ListModel::populateState() against each model's filter_fields, and the
direction is held to ASC/DESC by the same method, so the escape() call
around either was inert. Switching the column to quoteName() and dropping
the redundant escape() around the direction lets each query state what
its parts are.

The issue enumerated 25 sites from the single-statement shape
order($db->escape(. A value-anchored sweep found 15 further models that
assemble the clause into a local first and pass it to order() two lines
later, which no order()-anchored pattern returns. Those are included
here, bringing the total to 44 sites across 42 models.

Each of the 42 was checked to define filter_fields and route through
parent::populateState(). OrdersModel additionally merges plugin-supplied
sort columns, already held to a bare or dotted identifier by
getPluginColumns(). No model was found without that resolution, so this
is a consistency change throughout. filter_fields entries that name a
SELECT alias rather than a table column (total_qty, order_count,
totalitems, remaining_balance, uses_count, derived_status) carry no
parentheses or spaces, so each quotes to a valid identifier;
ReportitemisedModel already used quoteName() for its secondary sort and
its primary sort now matches.

Product ids read with ?? 0 are cast before reaching int-typed calls under
strict_types=1 -- ProductHelper (feeding getTraits() and traitsBelongTo())
and J2commerceField (feeding calculateVariantStats() and getImageCount())
-- matching the cast already present on the third such line in
ProductHelper. The now-redundant cast at the getProductUrl() call site is
folded into its assignment.

+++ administrator/components/com_j2commerce/src/Helper/ProductHelper.php
@@ -2000,7 +2000,7 @@ class ProductHelper
     {
         static $cache = [];
 
-        $productId = $product->j2commerce_product_id ?? 0;
+        $productId = (int) ($product->j2commerce_product_id ?? 0);
 
         if (!$productId) {
             return [];
@@ -4257,8 +4257,8 @@ class ProductHelper
      */
     public function getProductLink(object &$product): void
     {
-        $productId             = $product->j2commerce_product_id ?? 0;
-        $product->product_link = J2CommerceHelper::platform()->getProductUrl(['task' => 'view','id' => (int) $productId]);
+        $productId             = (int) ($product->j2commerce_product_id ?? 0);
+        $product->product_link = J2CommerceHelper::platform()->getProductUrl(['task' => 'view','id' => $productId]);
     }
 
     /**
